Added an exponential decay factor
This commit introduces an exponential decay factor, whose decay
rate is gamma-distributed. It also adds a regression test using
a sum of a decay and two impulses. The test needs tuning.

While checking the 'gauss' test, I noticed that the optimizer was
performing moves that decreased the variational bound. There is
a bug somewhere!!!