Moved to 'extensible' model types
This is the first commit in a series of commits intended to
change the way VFL interacts with other code. In particular,
it enables fully programmatic control over type-specific
model allocation and usage.

The factor_t and optim_t codebases will be ported in a similar
manner. Finally, a master factory, probably called vfl_t, will
be created that handles flexible object allocation.

This extensibility and ease of programmatic access form the
foundation of a scripting engine for variational features.