Initial port of factor_t to extensibility
This commit does the same for factor_t as the previous two
commits did for model_t and optim_t.

Unfortunately, it breaks support for the fixed impulse and
product factors! This must be remedied by incorporating a
(programmatically accessible) methods table into the
factor_type_t structure. Factors that need custom
methods will use register them in the methods table,
which could be accessed by an as-yet unimplemented
interpreter.