Added support for writing kernel code
Ok, this needs explanation.

One as-yet unstated goal of this work is an efficient and
modular framework for Bayesian Active Learning, specifically
used with NMR signal models. This research is referred to as
Active Nonuniform Sampling, or "Active NUS".

In this active learning method, one searches the posterior
predictive variance of a GP trained on the current dataset
and parameterized by the factor covariances. However, this
search is computationally expensive. Thus, it was expected
that this task would be performed on one or more GPUs.

Herein lies the rub. VFL breaks almost every rule of GPU
programming design practices. How do we retain the flexibility
and modularity of VFL while supporting fast on-GPU variance
searches?

This commit answers that question. Factors are provided with
a kernel() function pointer, which generates GPU code for
computing its covariance kernel function. The model then
ties these strings of code together into a function that
can be shipped off to the OpenCL compiler at runtime.