Altered the scheme for mean-field updates
This commit changes the way that assumed-density mean-field
updates are performed in VFL. Rather than giving the entire
set of coefficients to a factor at once, the coefficients
corresponding to each observation are streamed to the
factor using multiple calls.

At the cost of a little more function call overhead, we can
save a ton of memory by adopting this "streaming update"
scheme.

I have yet to test this to make sure everything is working,
but a simple test should be coming soon using the cosine
factor and a synthetic dataset.