Moved to packed storage in search kernels
This commit attempts to solve two outstanding issues within
search_execute(). The first issue, that cholesky factorizations
were failing due to the use of single-precision array storage,
is addressed by using a double-precision matrix and routine
for decomposition. The second, that dense matrix storage and
traversal is just plain wasteful, was addressed by moving to
packed triangular storage on-device.