Began work to unify all types
[WARNING] This commit is completely broken!

This commit is an initial step towards an interpreted language.
While previous efforts to unify factors, models, and optimizers
have been useful, they haven't brought the code to a point
where a useful language could be dropped in from above.

This commit (and the entire branch, in fact) is devoted to
unifying all objects (including data_t, search_t, rng_t, etc.)
under a single object type system. The extended types (factors,
models, and optimizers) will remain, but as an extra hierarchy
on top of the base type.