Added std, nil, data methods
This commit adds the standard method library, which is just
an object where utility functions can be defined to use in
the language. Further, it adds a 'nil' object, which can
be returned as an 'empty object' (like Py_None) by
methods. Finally, it adds some properties and methods
to the dataset object, though the 'grid' property is
still not implemented.