Towards proper garbage collection
This commit begins the task of implementing simple
reference counting garbage collection into vflang,
a task which begins with a complete rewrite of the
tree evaluation code to yield a more principled,
disciplined interpreter.

The collector is not yet called, and reference
counting needs to be added into the factor,
model, and optim objects, but it's a start.