Fixed the GC bugfix, added model props
This commit fixes a fix that was applied to keep
the GC from collecting elements of transient
lists, but that was actually creating copies
of elements from non-transient lists too!

It also adds the 'wbar' and 'Sigma' properties
to the model structure, which enables getting
and setting the weight means and covariances
of variational feature models.